onprem-app/service-b/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
PyJWT==2.9.0
